+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export { isRequestUrlInAssets } from './isRequestUrlInAssets.js';
-export { matchByUrl } from './matchByUrl.js';
+export { matchByUrl, type MatchByUrlOptions } from './matchByUrl.js';
 export { normalizeUrl } from './normalizeUrl.js';
 export { notifyClients } from './notifyClients.js';
 export { resolveAssetUrls } from './resolveAssetUrls.js';